Removes redundant assignment of `attributes` to itself
In more detail, it assigns to `attributes` a new Object whose keys are
all the values of `sorting[i].value`.

These values should be all the possible attributes, so in effect, the
keys don't change.

The value assigned to each of these keys is `attributes[key]`, in other
words, the value which that key **already has** in `attributes`.

So, `attributes` should not change. So the line is redundant.

As it prevents `attributes` from being made `const`, I've removed it.