Check word boundary in front of attribute keys (#825)
Sleek incorrectly parsed keys ending in 't' or 'due' and incorrectly
rendered labels for attributes ending in other keys we use.

about:1234 => was parsed as threshold
foodue:1234 => was parsed as due date

foo+bar => rendered as a project label

This is fixed by checking for the word boundary in front of the key,
making sure no random other characters are directly before the key.

Fixes #794